requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto==2.49.0
boto3==1.36.10
botocore==1.36.10
Brotli==1.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
contourpy==1.3.0
crcmod==1.7
cryptography
cycler==0.12.1
decorator==4.4.2
deepface==0.0.93
distro==1.9.0
dlib==19.24.2
exceptiongroup==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.8
fasteners==0.19
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
gcs-oauth2-boto-plugin==3.2
gdown==5.2.0
grpcio
grpcio-status
gsutil==5.33
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
imutils==0.5.4
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.7
libclang==18.1.1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==10.6.0
moviepy==2.0.0.dev2
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.1.0
namex==0.0.8
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.63.2
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pip==25.0
proglog==0.1.10
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyOpenSSL==24.2.1
pyparsing==3.2.1
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyu2f==0.1.5
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
retry_decorator==1.1.1
rich==13.9.4
rsa==4.7.2
s3transfer==0.11.2
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
starlette==0.45.3
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==3.0.1
urllib3==1.26.20
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0